Fix extracting logic in dynamic filtering when integrated with filter pushdown
When integrating with filter pushdown, we extract dynamic filter
and leave it in the filter node and push the regular filters to table scan.
The current extracting logic can't handle conjunction dynamic filters
